requirements.txt
python-dotenv==0.21.0
selenium==4.7.2
webdriver_manager==3.8.5
